Updating docs and changing copy to move
Updating the README to further clarify the instructions for using
hot deploy with WildFly.

Changed the copy_artifacts function in all assemble scripts to
move_artifacts instead of keeping two copies of the artifact around
which could confuse users about where to rsync or scp their artifacts
to use the hot deploy feature.