Use Pod IP address for the public interface
Use `hostname -i` to determine the IP address of the Pod and bind the
WildFly public interface to it.

Do not bind jgroups.bind_address to 0.0.0.0 (that is incorrect for
JGroups).
Instead, if JGroups subsystem is defined, the standalone configuration
can attach its socket-binding to the public interface so that the Pod IP
addres is properly used to identiy JGroups nodes.